server/requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
Requests==2.31.0
SQLAlchemy==2.0.27
PyMySQL==1.1.0
gunicorn==21.2.0